Rename group to io.alex
