--- src/main/kotlin/com/wolt/wm/training/bank/BankApplication.kt
@@ -1,11 +1,0 @@
-package com.wolt.wm.training.bank
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class BankApplication
-
-fun main(args: Array<String>) {
-    runApplication<BankApplication>(*args)
-}
