--- src/main/kotlin/com/wolt/wm/training/bank/utils/ParseUuidFromString.kt
@@ -1,14 +1,0 @@
-package com.wolt.wm.training.bank.utils
-
-import java.util.UUID
-
-fun parseUuidFromString(
-    str: String,
-    errorMsg: String? = "Invalid UUID",
-): UUID {
-    return try {
-        UUID.fromString(str)
-    } catch (e: IllegalArgumentException) {
-        throw IllegalArgumentException(errorMsg)
-    }
-}
