+++ src/main/kotlin/io/alex/bank/BankApplication.kt
@@ -1,0 +1,11 @@
+package io.alex.bank
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class BankApplication
+
+fun main(args: Array<String>) {
+    runApplication<BankApplication>(*args)
+}
