--- src/test/kotlin/com/wolt/wm/training/bank/BankApplicationTests.kt
@@ -1,9 +1,0 @@
-package com.wolt.wm.training.bank
-
-import org.junit.jupiter.api.Test
-
-class BankApplicationTests : IntegrationBaseTest() {
-    @Test
-    fun contextLoads() {
-    }
-}
