+++ src/test/kotlin/io/alex/bank/BankApplicationTests.kt
@@ -1,0 +1,9 @@
+package io.alex.bank
+
+import org.junit.jupiter.api.Test
+
+class BankApplicationTests : IntegrationBaseTest() {
+    @Test
+    fun contextLoads() {
+    }
+}
